Refactor garbage collection from configuration reconciler
