+++ pkg/reconciler/configuration/configuration_test.go
@@ -18,7 +18,6 @@ package configuration
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
@@ -29,7 +28,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	clientgotesting "k8s.io/client-go/testing"
 	duckv1beta1 "knative.dev/pkg/apis/duck/v1beta1"
 	"knative.dev/pkg/configmap"
@@ -38,9 +36,7 @@ import (
 	"knative.dev/pkg/ptr"
 	"knative.dev/serving/pkg/apis/serving/v1alpha1"
 	"knative.dev/serving/pkg/apis/serving/v1beta1"
-	"knative.dev/serving/pkg/gc"
 	"knative.dev/serving/pkg/reconciler"
-	"knative.dev/serving/pkg/reconciler/configuration/config"
 	"knative.dev/serving/pkg/reconciler/configuration/resources"
 
 	. "knative.dev/pkg/reconciler/testing"
@@ -404,153 +400,6 @@ func TestReconcile(t *testing.T) {
 			Base:                reconciler.NewBase(ctx, controllerAgentName, cmw),
 			configurationLister: listers.GetConfigurationLister(),
 			revisionLister:      listers.GetRevisionLister(),
-			configStore: &testConfigStore{
-				config: ReconcilerTestConfig(),
-			},
-		}
-	}))
-}
-
-func TestGCReconcile(t *testing.T) {
-	now := time.Now()
-	tenMinutesAgo := now.Add(-10 * time.Minute)
-
-	old := now.Add(-11 * time.Minute)
-	older := now.Add(-12 * time.Minute)
-	oldest := now.Add(-13 * time.Minute)
-
-	table := TableTest{{
-		Name: "delete oldest, keep two",
-		Objects: []runtime.Object{
-			cfg("keep-two", "foo", 5556,
-				WithLatestCreated("5556"),
-				WithLatestReady("5556"),
-				WithObservedGen),
-			rev("keep-two", "foo", 5554, MarkRevisionReady,
-				WithRevName("5554"),
-				WithCreationTimestamp(oldest),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-two", "foo", 5555, MarkRevisionReady,
-				WithRevName("5555"),
-				WithCreationTimestamp(older),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-two", "foo", 5556, MarkRevisionReady,
-				WithRevName("5556"),
-				WithCreationTimestamp(old),
-				WithLastPinned(tenMinutesAgo)),
-		},
-		WantDeletes: []clientgotesting.DeleteActionImpl{{
-			ActionImpl: clientgotesting.ActionImpl{
-				Namespace: "foo",
-				Verb:      "delete",
-				Resource: schema.GroupVersionResource{
-					Group:    "serving.knative.dev",
-					Version:  "v1alpha1",
-					Resource: "revisions",
-				},
-			},
-			Name: "5554",
-		}},
-		Key: "foo/keep-two",
-	}, {
-		Name: "keep oldest when no lastPinned",
-		Objects: []runtime.Object{
-			cfg("keep-no-last-pinned", "foo", 5556,
-				WithLatestCreated("5556"),
-				WithLatestReady("5556"),
-				WithObservedGen),
-			// No lastPinned so we will keep this.
-			rev("keep-no-last-pinned", "foo", 5554, MarkRevisionReady,
-				WithRevName("5554"),
-				WithCreationTimestamp(oldest)),
-			rev("keep-no-last-pinned", "foo", 5555, MarkRevisionReady,
-				WithRevName("5555"),
-				WithCreationTimestamp(older),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-no-last-pinned", "foo", 5556, MarkRevisionReady,
-				WithRevName("5556"),
-				WithCreationTimestamp(old),
-				WithLastPinned(tenMinutesAgo)),
-		},
-		Key: "foo/keep-no-last-pinned",
-	}, {
-		Name: "keep recent lastPinned",
-		Objects: []runtime.Object{
-			cfg("keep-recent-last-pinned", "foo", 5556,
-				WithLatestCreated("5556"),
-				WithLatestReady("5556"),
-				WithObservedGen),
-			rev("keep-recent-last-pinned", "foo", 5554, MarkRevisionReady,
-				WithRevName("5554"),
-				WithCreationTimestamp(oldest),
-				// This is an indication that things are still routing here.
-				WithLastPinned(now)),
-			rev("keep-recent-last-pinned", "foo", 5555, MarkRevisionReady,
-				WithRevName("5555"),
-				WithCreationTimestamp(older),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-recent-last-pinned", "foo", 5556, MarkRevisionReady,
-				WithRevName("5556"),
-				WithCreationTimestamp(old),
-				WithLastPinned(tenMinutesAgo)),
-		},
-		Key: "foo/keep-recent-last-pinned",
-	}, {
-		Name: "keep LatestReadyRevision",
-		Objects: []runtime.Object{
-			// Create a revision where the LatestReady is 5554, but LatestCreated is 5556.
-			// We should keep LatestReady even if it is old.
-			cfg("keep-two", "foo", 5556,
-				WithLatestReady("5554"),
-				// This comes after 'WithLatestReady' so the
-				// Configuration's 'Ready' Status is 'Unknown'
-				WithLatestCreated("5556"),
-				WithObservedGen),
-			rev("keep-two", "foo", 5554, MarkRevisionReady,
-				WithRevName("5554"),
-				WithCreationTimestamp(oldest),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-two", "foo", 5555, // Not Ready
-				WithRevName("5555"),
-				WithCreationTimestamp(older),
-				WithLastPinned(tenMinutesAgo)),
-			rev("keep-two", "foo", 5556, // Not Ready
-				WithRevName("5556"),
-				WithCreationTimestamp(old),
-				WithLastPinned(tenMinutesAgo)),
-		},
-		Key: "foo/keep-two",
-	}, {
-		Name: "keep stale revision because of minimum generations",
-		Objects: []runtime.Object{
-			cfg("keep-all", "foo", 5554,
-				// Don't set the latest ready revision here
-				// since those by default are always retained
-				WithLatestCreated("keep-all"),
-				WithObservedGen),
-			rev("keep-all", "foo", 5554,
-				WithRevName("keep-all"),
-				WithCreationTimestamp(oldest),
-				WithLastPinned(tenMinutesAgo)),
-		},
-		Key: "foo/keep-all",
-	}}
-
-	defer logtesting.ClearAll()
-	table.Test(t, MakeFactory(func(ctx context.Context, listers *Listers, cmw configmap.Watcher) controller.Reconciler {
-		return &Reconciler{
-			Base:                reconciler.NewBase(ctx, controllerAgentName, cmw),
-			configurationLister: listers.GetConfigurationLister(),
-			revisionLister:      listers.GetRevisionLister(),
-			configStore: &testConfigStore{
-				config: &config.Config{
-					RevisionGC: &gc.Config{
-						StaleRevisionCreateDelay:        5 * time.Minute,
-						StaleRevisionTimeout:            5 * time.Minute,
-						StaleRevisionMinimumGenerations: 2,
-					},
-				},
-			},
 		}
 	}))
 }
@@ -582,144 +431,4 @@ func rev(name, namespace string, generation int64, ro ...RevisionOption) *v1alph
 		opt(r)
 	}
 	return r
-}
-
-type testConfigStore struct {
-	config *config.Config
-}
-
-func (t *testConfigStore) ToContext(ctx context.Context) context.Context {
-	return config.ToContext(ctx, t.config)
-}
-
-var _ reconciler.ConfigStore = (*testConfigStore)(nil)
-
-func ReconcilerTestConfig() *config.Config {
-	return &config.Config{
-		RevisionGC: &gc.Config{
-			StaleRevisionCreateDelay: 5 * time.Minute,
-			StaleRevisionTimeout:     5 * time.Minute,
-		},
-	}
-}
-
-func TestIsRevisionStale(t *testing.T) {
-	curTime := time.Now()
-	staleTime := curTime.Add(-10 * time.Minute)
-
-	tests := []struct {
-		name      string
-		rev       *v1alpha1.Revision
-		latestRev string
-		want      bool
-	}{{
-		name: "fresh revision that was never pinned",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(curTime),
-			},
-		},
-		want: false,
-	}, {
-		name: "stale revision that was never pinned w/ Ready status",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-			},
-			Status: v1alpha1.RevisionStatus{
-				Status: duckv1beta1.Status{
-					Conditions: duckv1beta1.Conditions{{
-						Type:   v1alpha1.RevisionConditionReady,
-						Status: "True",
-					}},
-				},
-			},
-		},
-		want: false,
-	}, {
-		name: "stale revision that was never pinned w/o Ready status",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-			},
-			Status: v1alpha1.RevisionStatus{
-				Status: duckv1beta1.Status{
-					Conditions: duckv1beta1.Conditions{{
-						Type:   v1alpha1.RevisionConditionReady,
-						Status: "Unknown",
-					}},
-				},
-			},
-		},
-		want: true,
-	}, {
-		name: "stale revision that was previously pinned",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", staleTime.Unix()),
-				},
-			},
-		},
-		want: true,
-	}, {
-		name: "fresh revision that was previously pinned",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", curTime.Unix()),
-				},
-			},
-		},
-		want: false,
-	}, {
-		name: "stale latest ready revision",
-		rev: &v1alpha1.Revision{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "myrev",
-				CreationTimestamp: metav1.NewTime(staleTime),
-				Annotations: map[string]string{
-					"serving.knative.dev/lastPinned": fmt.Sprintf("%d", staleTime.Unix()),
-				},
-			},
-		},
-		latestRev: "myrev",
-		want:      false,
-	}}
-
-	cfgStore := testConfigStore{
-		config: &config.Config{
-			RevisionGC: &gc.Config{
-				StaleRevisionCreateDelay:        5 * time.Minute,
-				StaleRevisionTimeout:            5 * time.Minute,
-				StaleRevisionMinimumGenerations: 2,
-			},
-		},
-	}
-	ctx := cfgStore.ToContext(context.Background())
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			cfg := &v1alpha1.Configuration{
-				Status: v1alpha1.ConfigurationStatus{
-					ConfigurationStatusFields: v1alpha1.ConfigurationStatusFields{
-						LatestReadyRevisionName: test.latestRev,
-					},
-				},
-			}
-
-			got := isRevisionStale(ctx, test.rev, cfg)
-
-			if got != test.want {
-				t.Errorf("IsRevisionStale want %v got %v", test.want, got)
-			}
-		})
-	}
 }
