+++ pkg/reconciler/gc/config/testdata/config-gc.yaml
